der: add macro for `APPLICATION`, `CONTEXT-SPECIFIC` and `PRIVATE` tags
